==> 0_example_recommender.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
%matplotlib inline

stu_vle = pd.read_csv('data/stvl_ccc14b.csv')
stu_as = pd.read_csv('data/stas_ccc14b.csv')
ass = pd.read_csv('data/ass_ccc14b.csv')

vle = pd.read_csv('data/vle_ccc14b.csv')

This is an early draft of a study recommender. This program:

* looks at past student performance, particularly how much time students spent on individual study materials and their ultimate scores on assessments
* builds utility matrices based on that information for efficient look up
* creates study material recommendations for current students based on their similarity to past students

# Test run - assessment 3

ass.head()

Looks like assessment 24287 is the third one. We will also see that the 'recommended date' is a rough estimate. Students appear free to take assessments on their own time.

third_as = stu_as[stu_as.id_assessment.isin([24287])]
third_as

stu_as[stu_as.id_assessment.isin([24287])]['date_submitted'].mean()

## Create utility matrices

# Add a col indicating what date that student submitted the third assignment
stu_vle['third_sub_date'] = stu_vle['id_student'].map(third_as.set_index('id_student')['date_submitted'])

stu_vle.head()

stu_vle.groupby(['id_student']).head()

stu_as['id_student'].nunique()

third_as['id_student'].nunique()

next_as = third_as.set_index('id_student')

Students with NaN dropped out before the third assignment.

stu_vle.head()

#Dropping rows of students who withdrew before third assessment, and rows of clicks occurring on or after date the student took third assessment
stu_vle = stu_vle.drop(stu_vle[(stu_vle.date >= stu_vle.third_sub_date) | (stu_vle.third_sub_date.isnull())].index)

stu_vle.head()

Now let's sum up all the clicks per vle site. So we have one row per student-site combo.

# Sum up all clicks per vle site. Makes one row per student-site combo with all clicks up until third assessment.
stu_vle_sum = stu_vle.groupby(['id_student', 'id_site']).sum().reset_index()

stu_vle_sum.head()

Creating first utility matrix, for vle site study materials.

stuvle_matrix = stu_vle_sum.pivot(index='id_student', columns='id_site', values='sum_click')

stuvle_matrix.fillna(0, inplace=True)

stuvle_matrix.head()

# Dropping rows that are not first or second assessments
stuas_prev = stu_as.drop(stu_as[(stu_as.id_assessment != 24286) & (stu_as.id_assessment != 24282)].index)

stuas_prev.tail()

## Make similarity matrix

from sklearn.metrics.pairwise import cosine_similarity

sim_vle_matrix = cosine_similarity(stuvle_matrix, stuvle_matrix)

sim_vle_matrix = pd.DataFrame(sim_vle_matrix, index=stuvle_matrix.index, columns=stuvle_matrix.index)

sim_vle_matrix.head()

## Calculate ranks

vle_items = stuvle_matrix.columns.tolist()

students = stuvle_matrix.index.tolist()

stuvle_matrix.head()

stuas_prev.head()

next_as.head()

#### Debugging

clicks = stuvle_matrix.at[129198, 729671]
clicks

vle_sim = sim_vle_matrix.at[28418, 129198]
vle_sim

past_as_avg = get_past_score_avg(729671) / 100
past_as_avg

next_as_score = next_as.at[129198, 'score'] / 100
next_as_score

Hmmm.. this student does not have the first two assignments, but has the third. He also has a lot of vle activity. (see next cells) I guess the assessments in this course are.... optional?

stuas_prev[stuas_prev['id_student'] == 129198]

stu_as[stu_as['id_student'] == 129198]

len(stu_vle[stu_vle['id_student'] == 129198])

next_as.at[559157, 'score'] / 100

try:
    next_as.at[559257, 'score']
except:
    print("yikes")

if len(stuas_prev[stuas_prev['id_student'] == 129198]):
    print("yes")
else:
    print("no")

#### The functions

def get_past_score_avg(stu):
    # check if student has any past assessments
    if len(stuas_prev[stuas_prev['id_student'] == stu]):
        num_past_as = len(stuas_prev[stuas_prev['id_student'] == stu])
        total_points = stuas_prev[stuas_prev['id_student'] == stu]['score'].sum()
        return (total_points/num_past_as) / 100
    else:
        return None

def get_next_as_score(stu):
    try:
        score = next_as.at[559257, 'score']
        return score / 100
    except:
        return None

def get_ranks(stu):
    ranks = {}
    for vle_item in vle_items:
        item_rank = 0
        for student in students:
            # skip if it's the same student we are calculating ranks for, not needed in production which would use a new student
            if stu == student:
                continue
            clicks = stuvle_matrix.at[student, vle_item]
            vle_sim = sim_vle_matrix.at[stu, student]
            past_as_avg = get_past_score_avg(student)
            next_as_score = get_next_as_score(student)
            sim_score = vle_sim
            if past_as_avg:
                sim_score *= past_as_avg
            if next_as_score:
                sim_score *= next_as_score
            rank_score = sim_score * (1 + clicks)
            item_rank += rank_score
        ranks[vle_item] = item_rank
    return ranks

def get_study_recs(stu, num=3):
    total_students = len(stuvle_matrix)
    ranks = get_ranks(stu)
    ranked = sorted(ranks.items(), key=lambda x: x[1], reverse=True)
    picks = ranked[:num]
    print("To improve your score on the next assessment, try studying: ")
    for i in range(len(picks)):
        site = picks[i][0]
        print("Site #: ", site) 
        rec_clicks = (picks[i][1] / total_students) - stuvle_matrix.at[stu, site] #subtract however many clicks student has already given site
        print("     about ", rec_clicks, " more clicks.")

# Testing input student

get_study_recs(28418)

In production, the output above would replace "site number" with the name of the assignment in question.
