==> 0_subsets.ipynb <==
import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
%matplotlib inline

stu_vle = pd.read_csv('data/studentVle.csv')
stu_as = pd.read_csv('data/studentAssessment.csv')
vle = pd.read_csv('data/vle.csv')
ass = pd.read_csv('data/assessments.csv')
cou = pd.read_csv('data/courses.csv')
stu_reg = pd.read_csv('data/studentRegistration.csv')
stu_info = pd.read_csv('data/studentInfo.csv')

# Breaking out and saving single course subset of data

stvl_ = stu_vle[stu_vle.code_module.isin(['CCC'])][stu_vle.code_presentation.isin(['2014B'])]

asses = ass[ass.code_module.isin(['CCC'])][ass.code_presentation.isin(['2014B'])]['id_assessment'].tolist()
asses

stas_ = stu_as[stu_as.id_assessment.isin(asses)]

ass_ = ass[ass.code_module.isin(['CCC'])][ass.code_presentation.isin(['2014B'])]

ass_

vle_ = vle[vle.code_module.isin(['CCC'])][vle.code_presentation.isin(['2014B'])]

len(vle_)

stre_ = stu_reg[stu_reg.code_module.isin(['CCC'])][stu_reg.code_presentation.isin(['2014B'])]

stin_ = stu_info[stu_info.code_module.isin(['CCC'])][stu_info.code_presentation.isin(['2014B'])]

stvl_.to_csv('data/stvl_ccc14b.csv', index=False, header=True)
stas_.to_csv('data/stas_ccc14b.csv', index=False, header=True)
ass_.to_csv('data/ass_ccc14b.csv', index=False, header=True)

vle_.to_csv('data/vle_ccc14b.csv', index=False, header=True)

stre_.to_csv('data/stureg_ccc14b.csv', index=False, header=True)

stin_.to_csv('data/stuinfo_ccc14b.csv', index=False, header=True)
