==> 1_student_logreg.ipynb <==
# Logistic Regression

This model attempts to classify whether students ultimately succeed (pass) or not (fail/withdraw) in a course, based on student demographics information.

##### Summary:

Using only the most informative features did improve the score from 54% to 60%, but overall logistic regression based on student demographic information is not a very good model for predicting student success. However, this model's interpretability does give us some interesting takeaways in terms of which student features were selected:

*Not significant:*
* Age
* Gendere
* Disability status

*Significant:*
* Highest education level completed (positive)
* Number of previous attempts at the course (negative)
* Poverty level at place of origin (negative)

import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
%matplotlib inline

stu_info = pd.read_csv('data/studentInfo.csv')

stu_info.head()

## Preprocessing for Logistic Regression

stu_info.drop(['code_module', 'code_presentation'], axis=1, inplace=True)

stu_info.head()

stu_info.final_result.replace(to_replace=dict(Pass=1, Distinction=1, Fail=0, Withdrawn=0), inplace=True)

stu_info['gender'] = (stu_info['gender']=='M').astype(int)

stu_info['disability'] = (stu_info['disability']=='Y').astype(int)

stu_info.head()

stu_info['region'].unique()

stu_info['highest_education'].unique()

stu_info['imd_band'].unique()

stu_info['age_band'].unique()

#### Dummy variables for the categorical variables

cat_vars = ['region','highest_education','imd_band','age_band']
for var in cat_vars:
    cat_list = 'var' + '_' + var
    cat_list = pd.get_dummies(stu_info[var], prefix=var)
    stu_info1 = stu_info.join(cat_list)
    stu_info = stu_info1

stu_info.head()

info_vars = stu_info.columns.values.tolist()
to_keep = [i for i in info_vars if i not in cat_vars]

stuinfo_final = stu_info[to_keep]
stuinfo_final.columns.values

stuinfo_final.head()

stuinfo_final_vars = stuinfo_final.columns.values.tolist()
Y = ['final_result']
X = [i for i in stuinfo_final_vars if i not in Y ]

## Feature Selection

from sklearn import datasets
from sklearn.feature_selection import RFE
from sklearn.linear_model import LogisticRegression

model = LogisticRegression()

rfe = RFE(model, 15)
rfe = rfe.fit(stuinfo_final[X], stuinfo_final[Y] )
print(rfe.support_)
print(rfe.ranking_)

select_cols = [col for i, col in enumerate(X) if rfe.support_[i] == True]

select_cols

## Logistic Regression

Y = stuinfo_final['final_result']

X = stuinfo_final.drop(['final_result'], axis=1)

import statsmodels.api as sm
logit_model=sm.Logit(Y,X)
result=logit_model.fit()
print(result.summary())

from sklearn import linear_model
clf = linear_model.LogisticRegression()
clf.fit(X, Y)

clf.score(X,Y)

#### With Select Features Only

X = stuinfo_final[select_cols]

import statsmodels.api as sm
logit_model=sm.Logit(Y,X)
result=logit_model.fit()
print(result.summary())

from sklearn import linear_model
clf = linear_model.LogisticRegression()
clf.fit(X, Y)

clf.score(X,Y)
