==> 3_timeseries_features_model.ipynb <==
# Random Forest with TS-Fresh Feature Selection

The tsfresh python package is designed specifically for selecting features from time series data. With the help of having the best, most useful features, a random forest classifier achieves **79% accuracy** in predicting which students will ultimately succeed or fail based on the data from the first quarter of course timelines. Importantly, it also achieves **87% precision on class 0 (failing students)**. Depending on the nature of the interjection that this model would support, it may be worth it in a real-world setting to sacrifice precision in order to improve recall on class 0. (For example, if the interjection was polite and optional, it would probably be more important to reach students in danger of failing than to not bother students that are actually fine.)

import numpy as np
import pandas as pd

import matplotlib.pyplot as plt
%matplotlib inline

from tsfresh import extract_features, select_features, extract_relevant_features
from tsfresh.utilities.dataframe_functions import impute
from tsfresh.feature_extraction.settings import MinimalFCParameters, ComprehensiveFCParameters
from tsfresh.transformers import RelevantFeatureAugmenter

from sklearn.metrics import classification_report
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import Pipeline
import xgboost as xgb

stu_vle = pd.read_csv('data/stvl_ccc14b.csv')
stu_as = pd.read_csv('data/stas_ccc14b.csv')
ass = pd.read_csv('data/ass_ccc14b.csv')
vle = pd.read_csv('data/vle_ccc14b.csv')
stu_info = pd.read_csv('data/stuinfo_ccc14b.csv')
stu_reg = pd.read_csv('data/stureg_ccc14b.csv')

# Preprocessing

## NOTE: This is a demonstration of the preprocessing that needs to be done for using ts-fresh transformer. All of this preprocessing in encapsulated in the ts_preprocess function in src.util, for easier use of the pipeline in the future.

### Drop early withdrawal students

Or rather, determine which students to keep.

stu_reg.head()

pop_of_interest = stu_reg.drop(stu_reg[stu_reg.date_unregistration <= 67].index)['id_student'].unique()

len(stu_vle.drop(stu_vle[~stu_vle.id_student.isin(pop_of_interest)].index)['id_student'].unique()) # num students left in course who clicked

clkd_pop_of_interest = stu_vle.drop(stu_vle[~stu_vle.id_student.isin(pop_of_interest)].index)['id_student'].unique()

### Make y (targets) column

stu_info.head()

y_col = stu_info.drop(stu_info[~stu_info.id_student.isin(clkd_pop_of_interest)].index)

y_col = y_col.drop(['code_module', 'code_presentation', 'gender', 'region', 'highest_education', 'imd_band', 'age_band', 'num_of_prev_attempts', 'studied_credits', 'disability'], axis=1)
y_col.final_result.replace(to_replace=dict(Pass=1, Distinction=1, Fail=0, Withdrawn=0), inplace=True)
y_col = y_col.sort_values(by=['id_student'])

y_col.head()

y_col = y_col.set_index('id_student')

y = y_col['final_result']

### Train-test split BEFORE feature-extraction

Very important or else this would perform significantly worse in production.

y_train, y_test = train_test_split(y, test_size=0.2)

len(y_train), len(y_test)

### Make students click stream data (from which to generate features)

df_train = stu_vle.loc[stu_vle.id_student.isin(y_train.index)]
df_test = stu_vle.loc[stu_vle.id_student.isin(y_test.index)]

df_train = df_train.drop(['code_module', 'code_presentation', 'id_site'], axis=1)
df_train = df_train.groupby(['id_student', 'date']).sum() #sum daily clicks per student
df_train = df_train.reset_index()

df_test = df_test.drop(['code_module', 'code_presentation', 'id_site'], axis=1)
df_test = df_test.groupby(['id_student', 'date']).sum() #sum daily clicks per student
df_test = df_test.reset_index()

df_train.shape, df_test.shape

### Make X input for tsfresh feature augmenter (just student IDs of each group)

X_train = pd.DataFrame(index=y_train.index)
X_test = pd.DataFrame(index=y_test.index)

X_train.head()

# Create and run pipeline with feature extraction and classifier

This all may look strange but setting the parameters like this is precisely how the ts docs outline, for the use case of extracting features on the train set, and then only using those features when evaluating the test set. (that is, NOT extracting features from the test set) (see docs at https://github.com/blue-yonder/tsfresh/blob/master/notebooks/pipeline_with_two_datasets.ipynb)

ppl = Pipeline([('fresh', RelevantFeatureAugmenter(column_id='id_student', column_sort='date', 
                                                   default_fc_parameters=MinimalFCParameters())),
                ('clf', RandomForestClassifier())])

# for the fit on the train test set, we set the fresh__timeseries_container to `df_train`
ppl.set_params(fresh__timeseries_container=df_train)
ppl.fit(X_train, y_train)

# for the predict on the test test set, we set the fresh__timeseries_container to `df_test`
ppl.set_params(fresh__timeseries_container=df_test)
y_pred = ppl.predict(X_test)

print(classification_report(y_test, y_pred))

# Again but with diffs

df_trainpiv = df_train.pivot(index='id_student', columns='date', values='sum_click')
df_trainpiv.head()

df_trainpiv = df_trainpiv.fillna(0)

df_trpivdif = df_trainpiv.diff(axis=1)
df_trpivdif.drop(-18, axis=1, inplace=True) #drop first column which becomes all NaNs after diff

df_trpivdif.head()

df_trpivdif = df_trpivdif.stack() #back to tsfresh format

df_trpivdif = df_trpivdif.reset_index()

df_testpiv = df_test.pivot(index='id_student', columns='date', values='sum_click')
df_testpiv = df_testpiv.fillna(0)
df_tspivdif = df_testpiv.diff(axis=1)
df_tspivdif.drop(-18, axis=1, inplace=True) #drop first column which becomes all NaNs after diff
df_tspivdif = df_tspivdif.stack() #back to tsfresh format
df_tspivdif = df_tspivdif.reset_index()

df_tspivdif.head()

### pipeline

ppl2 = Pipeline([('fresh', RelevantFeatureAugmenter(column_id='id_student', column_sort='date', 
                                                   default_fc_parameters=MinimalFCParameters())),
                ('clf', RandomForestClassifier())])

# for the fit on the train test set, we set the fresh__timeseries_container to `df_train`
ppl2.set_params(fresh__timeseries_container=df_trpivdif)
ppl2.fit(X_train, y_train)

# for the predict on the test test set, we set the fresh__timeseries_container to `df_test`
ppl2.set_params(fresh__timeseries_container=df_tspivdif)
y_pred2 = ppl2.predict(X_test)

print(classification_report(y_test, y_pred2))

Much worse with diffs this time!

# Parameter play

ppl3 = Pipeline([('fresh', RelevantFeatureAugmenter(column_id='id_student', column_sort='date', 
                                                   default_fc_parameters=ComprehensiveFCParameters())),
                ('clf', RandomForestClassifier(n_estimators = 200,
                                max_depth=5,
                                random_state = 21))])

# for the fit on the train test set, we set the fresh__timeseries_container to `df_train`
ppl3.set_params(fresh__timeseries_container=df_train)
ppl3.fit(X_train, y_train)

# for the predict on the test test set, we set the fresh__timeseries_container to `df_test`
ppl3.set_params(fresh__timeseries_container=df_test)
y_pred3 = ppl3.predict(X_test)

print(classification_report(y_test, y_pred3))

**That's more like it.** This model achieved 79% accuracy, and 87% precision on class 0 (failing students). Depending on the nature of the interjection that this model would support, it may be worth it in a real-world setting to sacrifice precision to improve recall on class 0.

ppl3.named_steps['clf'].feature_importances_

ppl4 = Pipeline([('fresh', RelevantFeatureAugmenter(column_id='id_student', column_sort='date', 
                                                   default_fc_parameters=MinimalFCParameters())),
                ('clf', xgb.XGBClassifier(n_estimators    = 300, 
                              max_depth    =  5,
                              random_state = 21))])

# for the fit on the train test set, we set the fresh__timeseries_container to `df_train`
ppl4.set_params(fresh__timeseries_container=df_train)
ppl4.fit(X_train, y_train)

# for the predict on the test test set, we set the fresh__timeseries_container to `df_test`
ppl4.set_params(fresh__timeseries_container=df_test)
y_pred4 = ppl4.predict(X_test)

print(classification_report(y_test, y_pred4))
